requirements/base.txt
argon2-cffi==23.1.0
django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-spectacular==0.27.2
celery==5.4.0
django-celery-beat==2.7.0
django-celery-email==3.0.0
flower==2.0.1
redis==5.0.8
pytz==2024.1
